+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2026.6.1
+
+- Added more SES settings
+- Updated the auto-updater
+- Added the ability to switch the tree view between chains and categories
+- Added new carbon-based color schemes
+
 ## 2026.6.0
 
 - New VTX beta